+++ src/lib.rs
@@ -79,7 +79,6 @@ pub mod stats;
 pub mod subcommand_compose;
 pub mod subcommand_inspect;
 pub mod subcommand_list_codecs;
-pub mod subcommand_rtmp_outbound_endpoint;
 pub mod subcommand_server;
 pub mod subcommand_tune;
 pub mod subcommand_vmaf;